refactor: move hex editor to edit directory
