+++ Assets/Editor/HexCoordinatesDrawer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: efbd0cb988412644fa5ca2d64b6b9089
+guid: bef7e3bdfa06fa347a4ef4a9ba5cecd6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
