--- Assets/Scripts/Hex/Editor/HexCoordinatesDrawer.cs
@@ -1,20 +1,0 @@
-﻿using UnityEngine;
-using UnityEditor;
-
-namespace Hex
-{
-    [CustomPropertyDrawer(typeof(HexCoordinates))]
-    public class HexCoordinatesDrawer : PropertyDrawer
-    {
-        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
-        {
-            HexCoordinates coordinates = new HexCoordinates(
-                property.FindPropertyRelative("x").intValue,
-                property.FindPropertyRelative("z").intValue
-            );
-
-            position = EditorGUI.PrefixLabel(position, label);
-            GUI.Label(position, coordinates.ToString());
-        }
-    }
-}
